+++ chaos_theory/src/lib.rs
@@ -124,6 +124,8 @@ mod make_combine;
 mod make_core;
 #[cfg(feature = "ecow")]
 mod make_ecow;
+#[cfg(feature = "either")]
+mod make_either;
 mod make_float;
 #[cfg(feature = "hashbrown")]
 mod make_hashbrown;
@@ -212,6 +214,12 @@ pub mod make {
     /// [`ecow`](https://docs.rs/ecow) generators.
     pub mod ecow {
         pub use crate::make_ecow::*;
+    }
+
+    #[cfg(feature = "either")]
+    /// [`either`](https://docs.rs/either) generators.
+    pub mod either {
+        pub use crate::make_either::*;
     }
 
     #[cfg(feature = "indexmap")]